requirements.txt
streamlit==1.50.0
pandas==2.2.3
plotly==5.18.0
psycopg2-binary==2.9.10
bcrypt==4.2.0
pyyaml==6.0.3
streamlit-authenticator==0.4.2
